Cleaned up little bit of code
