+++ css/main.css
@@ -117,7 +117,7 @@ input[type="file"] {
 }
 
 .map-color{
-  background-color: #491a90;
+  background-color: #a4a02d;
 }
 
 .button:hover {
