+++ src/map.js
@@ -64,4 +64,14 @@ function normalizeList(list) {
   return list;
 }
 
-export { parseMapFromSocket, parseMapFromLcm, normalizeList };
+function downloadObjectAsJson(exportObj, exportName){
+  var dataStr = "data:application/json;charset=utf-8," + encodeURIComponent(JSON.stringify(exportObj));
+  var downloadAnchorNode = document.createElement('a');
+  downloadAnchorNode.setAttribute("href",     dataStr);
+  downloadAnchorNode.setAttribute("download", exportName + ".json");
+  document.body.appendChild(downloadAnchorNode); // required for firefox
+  downloadAnchorNode.click();
+  downloadAnchorNode.remove();
+}
+
+export { parseMapFromSocket, parseMapFromLcm, normalizeList, downloadObjectAsJson };
